refactor: move serializers into dedicated test helpers directory
