+++ src/scripts/__tests__/format.js
@@ -1,5 +1,8 @@
 import cases from 'jest-in-case'
-import {unquoteSerializer, winPathSerializer} from './helpers/serializers'
+import {
+  unquoteSerializer,
+  winPathSerializer,
+} from '../../test/helpers/serializers'
 
 expect.addSnapshotSerializer(unquoteSerializer)
 expect.addSnapshotSerializer(winPathSerializer)
