+++ src/test/helpers/serializers.js
@@ -21,5 +21,6 @@ function normalizePaths(value) {
   if (typeof value !== 'string') {
     return value
   }
+
   return slash(value.split(process.cwd()).join('<PROJECT_ROOT>'))
 }
